style: move custom exception to self-folder
