+++ src/main/java/org/sky/fly/handler/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package org.sky.fly.handler;
 
 import lombok.extern.slf4j.Slf4j;
-import org.sky.fly.exception.CustomException;
+import org.sky.fly.common.CustomException;
 import org.sky.fly.common.Result;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
